Conflicts:
	posts/views.py

comit culpa del dr hous ramirez

+++ posts/views.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
-from django.http.response import HttpResponse
 from django.shortcuts import render, get_object_or_404
 from posts.models import Post
+from socials.models import Puntuacio
 from posts.forms import PostForm, FiltreRutaForm
-from django.http.response import HttpResponseRedirect
+from django.http.response import HttpResponseRedirect, HttpResponse, Http404
 from django.contrib import messages
 from django.core.urlresolvers import reverse
 from django.contrib.auth.decorators import login_required
@@ -12,11 +12,18 @@ import math
 from django.core.paginator import Paginator, PageNotAnInteger, EmptyPage
 import json
 from django.utils.datetime_safe import datetime, date
+from django.views.decorators.http import condition
+from django.contrib.auth.decorators import permission_required
 
-def mostrarRutes(request):
-    q_admin = Q( administrador = request.user.perfil )
-    q_apuntat = Q( apuntats = request.user.perfil ) 
-    Rutes = Post.objects.filter( q_admin | q_apuntat )    
+@login_required
+def participaRuta(request):
+    Rutes = Post.objects.filter( apuntats = request.user.perfil )
+    return render(request, 'posts/mostrarRutes.html', {'Rutes':Rutes})
+
+
+@login_required
+def mevaRuta(request):
+    Rutes = Post.objects.filter(administrador=request.user.perfil)
     return render(request, 'posts/mostrarRutes.html', {'Rutes':Rutes})
 
 def mostrarRutesAcabades(request):    
@@ -39,8 +46,8 @@ def calcularDistanciaMapa(frase):
                 listCoordenades.append(coordenada1)
                 listCoordenades.append(coordenada2)
             # calculo la distancia    
-            i =0
-            distancia=0
+            i = 0
+            distancia = 0
             while i < len(listCoordenades)-3:
                 x1 = float(listCoordenades[i])
                 i=i+1
@@ -68,14 +75,19 @@ def calcularDuradaMapa(modo, distancia):
             
             return tiempo
 
-
+@login_required
 def eliminarRuta(request, ruta_id):
     ruta=get_object_or_404(Post, pk=ruta_id)
+    #seguretat
+    if ruta.administrador != request.user.perfil:
+            url_next= reverse('index', kwargs={})
+            return HttpResponseRedirect(url_next)
     ruta.delete()
     
     url_next= reverse('posts:mostrarRutes', kwargs={})
     return HttpResponseRedirect(url_next)
 
+@login_required
 def apuntarRuta(request, ruta_id):
     ruta = get_object_or_404(Post, pk=ruta_id)
     
@@ -87,6 +99,7 @@ def apuntarRuta(request, ruta_id):
         
     url_next= reverse('posts:mostrarRutes', kwargs={})
     return HttpResponseRedirect(url_next)
+
 
 def desapuntarRuta(request, ruta_id):
     ruta = get_object_or_404(Post, pk=ruta_id)
@@ -103,6 +116,11 @@ def editaRuta(request, ruta_id=None):
 
     if ruta_id is not None:
         ruta=get_object_or_404(Post, pk=ruta_id)
+        #seguretat
+        if ruta.administrador != request.user.perfil:
+            messages.add_message(request, messages.ERROR , "No pots modificar aquesta ruta!")
+            url_next= reverse('index', kwargs={})
+            return HttpResponseRedirect(url_next)
     else:
         ruta=Post()
         
@@ -138,7 +156,9 @@ def editaRuta(request, ruta_id=None):
         'form':form,
         })
 
-
+def detall_ruta(request, ruta_id):
+    ruta = Post.objects.filter(pk = ruta_id)
+    return render(request, 'posts/detall.html', {'ruta':ruta})
 
 def filtreDeRutes(request):
 
@@ -191,8 +211,35 @@ def filtreDeRutes(request):
     return render(request, 'posts/filtreDeRutes.html', {'form':form, 'rutes':rutes, 'q':q_str})
 
 
+def puntuar(request):
+    punts = request.GET.get('punts')
+    ruta_id = request.GET.get('ruta_id')
+    
+    
+    usuari = request.user.perfil
+    
+    q = Q()
+    q = Q(apuntats = usuari)
+    q &= Q(pk = ruta_id)
+    buscador = Post.objects.filter(q).first()
+    
+    if buscador is not None:
+        p = Puntuacio.objects.get( q );
+            
+    else:
+        p = Puntuacio()
+        p.perfil = usuari
+        ruta = Post.objects.get(pk = ruta_id)
+        p.post = ruta
+    
+    p.puntuacio = punts
+    p.save()
+    
+    return HttpResponse(content='OK')
+
+
 def paginaitor_plus(page, llista, num):
-    paginator = Paginator(llista, num) # numero d'entrades per pagina
+    paginator = Paginator(llista, num) # numero d'entrades per pàgina
     try:
         entrada = paginator.page(page)
     except PageNotAnInteger:
